test(flags): Mutualize tests
